[F] Fix user graduation year type error

+++ src/main/java/org/hydev/veracross/analyzer/database/model/User.java
@@ -72,7 +72,7 @@ public class User
 
     @NonNull
     @Column(name = "graduation_year", length = 6)
-    private Integer graduationYear;
+    private Short graduationYear;
 
     @NonNull
     @Column(name = "groups")
